+++ lib/atoms/text_controller/initial_text_controller.dart
@@ -6,16 +6,36 @@ class InitialTextController extends TextEditingController {
   final String fixedInitialString;
   final TextStyle? fixedInitialStringStyle;
 
-  InitialTextController({String? text, required this.fixedInitialString, this.fixedInitialStringStyle}) : super(text: '$fixedInitialString ${text ?? ''}');
+  InitialTextController({String? text, required this.fixedInitialString, this.fixedInitialStringStyle})
+      : super(text: '$fixedInitialString${text ?? ''}') {
+    super.value = value.copyWith(selection: TextSelection.collapsed(offset: fixedInitialString.length));
+  }
 
   @override
   set value(TextEditingValue newValue) {
-    if (newValue.text.length < fixedInitialString.length) return;
-    if (newValue.selection.baseOffset < fixedInitialString.length + 1)
-      newValue = newValue.copyWith(selection: TextSelection(baseOffset: fixedInitialString.length + 1, extentOffset: max(newValue.selection.extentOffset, fixedInitialString.length + 1)));
+    if (newValue.text.isEmpty) newValue = newValue.copyWith(text: fixedInitialString);
 
-    if (newValue.selection.baseOffset > newValue.selection.extentOffset && newValue.selection.extentOffset < fixedInitialString.length + 1)
-      newValue = newValue.copyWith(selection: TextSelection(baseOffset: newValue.selection.baseOffset, extentOffset: max(newValue.selection.extentOffset, fixedInitialString.length + 1)));
+    if (newValue.text.length < fixedInitialString.length) return;
+    if (newValue.selection.baseOffset < fixedInitialString.length)
+      newValue = newValue.copyWith(
+        selection: TextSelection(
+          baseOffset: fixedInitialString.length,
+          extentOffset: max(newValue.selection.extentOffset, fixedInitialString.length),
+        ),
+      );
+
+    if (newValue.selection.baseOffset > newValue.selection.extentOffset && newValue.selection.extentOffset < fixedInitialString.length)
+      newValue = newValue.copyWith(
+        selection: TextSelection(
+          baseOffset: newValue.selection.baseOffset,
+          extentOffset: max(newValue.selection.extentOffset, fixedInitialString.length),
+        ),
+      );
+
+    if (newValue.text.substring(0, fixedInitialString.length) != fixedInitialString)
+      newValue = newValue.copyWith(
+        text: text.replaceRange(0, fixedInitialString.length, fixedInitialString),
+      );
 
     super.value = newValue;
   }